Review fixes: dedupe normalizeCdpUrl, rename page-state to page-utils, tighten exports
